[init] Accelerator installed globally
Preference is that the accelerator runs from a global directory.
All references to HOME are avoided and replaced by references
under somewhere in '/'. This allows for different users to access
the accelerator subsystem, without requiring access to a specific
user home folder. Access can then be determined using permissions
and ACL if applicable